+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Add class `BWidgets::Activatable` and implement in `BWidgets::Widget`
 * Add class `BWidgets::Navigatable` and implement in `BWidgets::Box`
 * Fix prevent twice addition of events to the event loop
+* Add `BEvents::Event::EventType::all`
 * Update documentation